Some tests:

```
$ aux="../ex_doc/bin/ex_doc ExUnit 1.0.5 lib/ex_unit/ebin -u https://github.com/elixir-lang/elixir -o doc/ex_unit -p http://elixir-lang.org/docs.html"

$ # current version
$ eval "time $aux"
../ex_doc/bin/ex_doc ExUnit 1.0.5 lib/ex_unit/ebin -u  -o doc/ex_unit -p   0.67s user 0.27s system 125% cpu 0.751 total

$ # new version, sync mode (-m given)
$ eval "time $aux -m ExUnit"
../ex_doc/bin/ex_doc ExUnit 1.0.5 lib/ex_unit/ebin -u  -o doc/ex_unit -p  -m   0.67s user 0.25s system 135% cpu 0.673 total

$ # new version, async mode (without entry-point module in docs)
$ eval "time $aux"
./ex_doc/bin/ex_doc ExUnit 1.0.5 lib/ex_unit/ebin -u  -o doc/ex_unit -p   0.52s user 0.23s system 123% cpu 0.604 total
```