@ericmj 2015-06-17
Please take a look at the `Task` module, `async`/`await` in particular, instead of sending/receiving manually.