@milmazz 2015-06-18
@ericmj Thanks for your feedback. Please let me know if the last commit fulfill your expectations.